refactor: Extract and reorganize type definitions into dedicated `ui`, `transfer`, and `receiver` type files.
