+++ web-app/src/types/ui.ts
@@ -1,0 +1,10 @@
+export interface AlertDialogState {
+    isOpen: boolean
+    title: string
+    description: string
+    type: 'success' | 'error' | 'info'
+}
+
+export type AlertType = 'success' | 'error' | 'info'
+
+export type InstructionsCardProps = {}
